Fix helm arg processing of empty strings
We were misusing the IntOrStr type and just blindly assuming that
anything with an empty StrVal was an integer, when there is a specific
Type field to tell us if it is an integer or not. We were also less
flexible than Helm when handling booleans, and didn't handle null
properly either.

Signed-off-by: Brad Davidson <brad.davidson@rancher.com>