Avoid hard-coded values for API server address
Avoid hard-coded value of 127.0.0.1 and 6443, as the API server can
listen on another address when used outside of k3s.

When used from another go program using the `chart` package, host and
ports should be passed as argument.

When used standalone, host and port should be derived from the
kubeconfig and/or master-url options.